Specify `auth_scheme: :request_body` for oauth2 v2.x compatibility
The oauth2 gem switched the default auth_scheme in v2 from `:request_body`
to `:basic_auth`. This caused an "Invalid Credentials" error when
requesting the auth token because client_id and client_secret params
were not included in the requests.

This is a backwards compatible change that will work with oauth2 v1.x and v2.x.

https://gitlab.com/oauth-xx/oauth2/-/blob/v1.4.11/lib/oauth2/client.rb?ref_type=tags#L44
https://gitlab.com/oauth-xx/oauth2/-/blob/v2.0.0/lib/oauth2/client.rb?ref_type=tags#L46